fix: address PRRT_kwDOSJAM6s6K0ODk - split oversized executor coverage shard
