+++ plans/PRRT_kwDOSJAM6s6K0ODk_VALIDATION.md
@@ -1,0 +1,26 @@
+# PRRT_kwDOSJAM6s6K0ODk Validation
+
+## Result
+
+The review feedback was actionable. The terminal conformance failure regression
+was moved from part 001 into a new executor coverage shard, reducing
+`test_executor_coverage_edges_part_001.py` below the maintainability line limit
+without changing assertions or production behavior.
+
+## Evidence
+
+- Before change:
+  - `test_executor_coverage_edges_part_001.py`: 1,510 lines.
+- After change:
+  - `test_executor_coverage_edges_part_001.py`: 1,348 lines.
+  - `test_executor_coverage_edges_part_012.py`: 187 lines.
+- Focused checks:
+  - `uv run --python 3.12 --extra dev ruff check tests/unit/control/test_executor_coverage_edges_parts/test_executor_coverage_edges_part_001.py tests/unit/control/test_executor_coverage_edges_parts/test_executor_coverage_edges_part_012.py`
+  - `uv run --python 3.12 --extra dev pytest tests/unit/control/test_executor_coverage_edges_parts/test_executor_coverage_edges_part_012.py::test_validation_conformance_failure_still_deposits_before_mark_failed -q`
+  - direct line-count assertion for parts 001 and 012 against the 1,500-line limit.
+
+## Notes
+
+Full AWF/GitHub validation is managed after agent completion. I did not run the
+full maintainability suite because `test_executor_coverage_edges_part_002.py`
+is an existing unrelated oversized shard covered by separate review feedback.
